towards-targets: Move operator modules to separate directory
