+++ devito/__init__.py
@@ -22,7 +22,7 @@ from devito.compiler import compiler_registry
 from devito.dle import dle_registry
 from devito.dse import dse_registry
 from devito.logger import error, warning, info, logger_registry, set_log_level  # noqa
-from devito.profiling import profiler_registry
+from devito.operator import profiler_registry
 
 
 from ._version import get_versions  # noqa
