+++ devito/operator/__init__.py
@@ -1,0 +1,2 @@
+from .operator import Operator  # noqa
+from .profiling import profiler_registry  # noqa
